+++ CHANGELOG.md
@@ -1,4 +1,8 @@
-# Changes to QSA Scope
+# Changes to Element QSA Scope
+
+### 1.1.0 (July 5, 2018)
+
+- Polyfill `matches` and `closest`
 
 ### 1.0.1 (January 24, 2017)